[Feat] 임장노트 나누기 선택 작업
