+++ juinjang/Projects/App/Sources/UserInterface/ImjangList/ImjangListViewController.swift
@@ -169,7 +169,7 @@ extension ImjangListViewController: DeleteImjangListDelegate {
     
     // 임장노트 나누기 화면으로 이동
     @objc private func showDivideImjangVC() {
-        let shareImjangVC = ShareImjangNoteViewController()
+        let shareImjangVC = ShareImjangNoteViewController(rootView: ShareImjangNoteView(frame: .zero))
         let repository = MockShareImjangNoteRepository()
         shareImjangVC.reactor = ShareImjangNoteReactor(repository: repository)
         self.navigationController?.pushViewController(shareImjangVC, animated: true)
